优化

+++ ETMS.Entity/Database/Source/EtActivityRoute.cs
@@ -90,7 +90,7 @@ namespace ETMS.Entity.Database.Source
         /// <summary>
         /// <see cref="EmActivityRouteStatus"/>
         /// </summary>
-        public int RouteStatus { get; set; }
+        public byte RouteStatus { get; set; }
 
         public string ShareQRCode { get; set; }
